Implement trace2 manager taking care of trace2 integration
This commit implement a manager object to manage trace2 integration to a
git command. This manager allows hook registration. Before the command
starts, it checks if any hook is interested in trace2 data. If any of
them does, it opens a temporary tempfile, injects the corresponding
environment variables to the command environment variable set. After the
command executes, it collects the result and parses it to the tree.
Finally, it dispatches activated hooks' handlers.