testdb: Pass database via parameter when terminating connections
The SQL statement to terminate database connections embeds the database
name that is to be terminated directly into the string. Even though we
know that the database name will always be a UUID that can be safely
passed, it still doesn't conform to best practices.

Refactor the code to instead use a proper parameter for the database
name.