datastore: Fix leaking Goroutine causing test flakiness
The common test logic for listener in `verifyListener()` is spawning a
Goroutine in which the listener is set up. This Goroutine eventually
gets cancelled via context cancellation so that it is unblockend and
exits. And in order to synchronize with it exiting, we wait for a
channel to be closed by that Goroutine.

While this works as expected in the general case, there is one early
exit in the helper in case the caller does not pass a verifier as input.
As a result it can happen that we exit the test while the Goroutine is
still running, and that can then race with the database being killed.
This causes the following flake:

    === FAIL: internal/praefect/datastore TestListener_Listen_storage_repositories_verification_updates_ignored (2.77s)
        listener_test.go:465:
                    Error Trace:	/builds/gitlab-org/gitaly/internal/praefect/datastore/listener_test.go:465
                                                            /builds/gitlab-org/gitaly/internal/praefect/datastore/asm_amd64.s:1594
                    Error:      	Should be true
                    Test:       	TestListener_Listen_storage_repositories_verification_updates_ignored
                    Messages:   	wait for notification: FATAL: terminating connection due to administrator command (SQLSTATE 57P01)

Fix this flake by deferring the context cancellation together with
waiting for the channel to be closed.