objectpool: Capture fsck stderr when fetching
We have observed the `git fsck` process run in `rescueDanglingObjects`
failing regularly. Sometimes this is because the process is taking
longer than 75 seconds and hitting its deadline, but many times we're
getting a return code. This tells us what went wrong generally, e.g. a
problem with the commit graph, refs, etc., but not the precise issue
with the repo.

To help understand what is causing these processes to fail, capture
`git fsck`'s stderr and include it with the error message.