Add some basic CLI tests for Gitaly
As we are about to migrate to start using a new CLI framework in
Gitaly, this commit adds some basic tests to ensure that the behavior
stays similar when we do the actual changes.

Version is included in the binaries built in tests to ensure the output
stays consistent with the main test binary.