Move Gitaly's CLI implementation to an internal package
The implementation of a binary's main function should be minimal and
should generally just call a library that actually implements it. This
ensures flexibility as you can run the binary instance as a library as
well. This commit moves Gitaly's main implementation and surrounding
code to the `internal/cli/gitaly` package. This keeps gives us a package
structure to use in the future and keeps the implementation itself internal
as well.