Do not use OS globals in `check` command
The `check` command currently has a few anti-patterns:

1. It's exiting directly with os.Exit. This is undesired as the
   program will exit direcly without chance to catch error on
   higher levels or run deferred functions. The command is mostly
   using os.Exit to exit with a specific error code. Replace these
   uses by returning an error which carries the desired exit code.
   This ensures the control flow is closer to the typical.
2. The command is accessing OS globals stdout and stderr directly.
   The command should instead access them through the context which
   ensures they can be substituted as necessary for example in tests.
   Relying on globals is bad in general, so let's use the injected
   outputs from the context.

Some of the called code is still using globals through a global logger
and environment variables. These are left for later as their uses span
a wider aread in the code base. The goal here is to align the command
to better patterns so we have a better standard to follow in future
command implementations.