gitaly: Fix transaction manager test flake
Currently the `propose returns if context is canceled before admission`
subtest in `TestTransactionManager` is flaky due to a race between
context cancellation and tranaction admission into the manager. To
resolve this, `SkipStartManager` is added to the step configuration to
run the subtest without starting the transaction manager. This prevents
transactions from being admitted to the manager and requires the context
to be cancelled before the test continues. With the race condtion
resolved, the quarantine is removed from the test.

Also the `StartManager` field is renamed to `RestartManager` to improve
the readability of the test since it more accurately reflects what is
happening during test execution.