Add stripPrefix option to ExtractHooks
The hooks come packaged in a TAR inside custom_hooks directory. Currently
ExtractHooks extracts this directory directly into the target folder. In
the upcoming WAL implementation for custom hooks, the hooks will be stored
in `wal/hooks/<log_index>`. That directory should contain the hooks directly
and it's unnecessary to put them in a custom_hooks subdirectory in there.

This commit makes this possible by adding the stripPrefix option for
ExtractHooks to remove the custom_hooks prefix from the extracted
hooks.