Add multiple finalizers support to internal/command
When spawning a command, we support adding a finalizer option via
`command.WithFinalizer`.  Originally, there is only one finalizer. To
prepare for trace2 integration, Git command factory needs to find a way
to hook to when after a command finishes. Finalizer is a good way to
achieve this purpose. Unfortunately, it's not easy to create a wrapper
function then trigger the pre-existing finalizer if any. So, expanding
this option to support multiple finalizers is a good approach.

This commit also adds context to the function signature of finalizer.