trace2: Implement trace2 hook that export events to distributed tracing
This hook receives the tree from the parser. Fortunately, the parsed
tree can be translated 1:1 to corresponding distributed tracing. The
span maintains the nesting structure and copied most of the data from
trace. Resulting spans are connected to the pre-existing parent span via
input context. This context is passed in when Git command finishes.