housekeeping: Refactor CleanStaleData tests to be less repetitive
In our tests for CleanStaleData we're using a bunch of helpers to create
a dummy filesystem hierarchy. Constructing the filesystem hierarchy is
quite repetitive though as tests need to always specify the name, mode,
age and expectations for every single entry even though most of the data
is the same across every entry.

Refactor the test to use sane defaults and options that allow the
testcases to override specific details for each of the file entries.
This makes the tests much more focussed on specifying what the actual
outliers are compared. Furthermore, it gives us an easy way to extend
the abstraction without having to adjust all testcases.