testhelper: Add `RequireTarState` for asserting tarball content state
Currently there is not a testhelper that assists with asserting the
state of a tarball's contents. Such a helper would make validating an
archive easier and eliminate the need to extract its contents to disk.
This change introduces `RequireTarState()`, which is used in future
commits, to validate the contents of a tarball.

`MustCreateCustomHooksTar()` is also added to aid in the creation of a
temporary custom hooks tarball to be used in testing.