testhelper: Clarify variable names in certificate generation
The variable names used by the function to generate certificates are a
bit confusing. Most importantly, using both "root" and "ca" for the
certificate authority makes one think that those are two different
things.

Rename the variables to clarify that there really only are two entities:
the "root" certificate that is supposed to sign the certificate of the
actual "entity".