Prepare TransactionManager tests for interface change
To provide support for snapshot reads, we need to change the
interface of the TransactionManager. Currently it only supports
queueing a transaction for committing. This doesn't allow for
explicitly opening a transaction, which is needed for snapshot reads
to determine the snapshot and to keep the versions alive until the
transaction finishes. To make the diff more manageable when we make
the interface change, this commit changes the tests to not access
the Transaction struct directly as the accesses will be later guarded
by methods.