Provide means to open a transaction through TransactionManager
TransactionManager currently doesn't support opening a transaction,
just queuing changes to be performed. This isn't sufficient, as we
need a way to open a transaction to ensure it's snapshot remains the
same throughout the transaction. This commit changes TransactionManager
to provide an explicit interface for managing a transaction similar to
other databases. Begin opens a transaction and the returned transaction
can be operated upon to stage changes to be ultimately committed or
rolled back. This commit updates the interface only but doesn't yet
do any synchronization otherwise.