praefect: Sub-command 'configuration validate'
In order to improve developers experience in using Praefect
the new 'configuration' sub-command is added. It has only one
sub-command 'validate' and the purpose of it is to validate provided
configuration before starting the service. So far it is only a
stub that does nothing, but introduces a cli command.

Changelog: added

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/4650