gittest: Support TCP and TLS connections when dialing server
The gittest package needs to connect to the gRPC server when either
creating repositories via the Repository service or when retrieving the
replica path of a repository. Both of these usecases are implemented via
`dialService()`, which right now only supports connecting to a server
that is listening on its socket path. This restricts the usefulness of
this option for tests which use a different server configuration.

Improve the logic to also support dialing of servers that listen on
either the TCP or the TLS socket. While at it, mark the function as a
test helper.