repository: Simplify CreateFork test for dialing TLS servers
One of the tests for CreateFork verifies that things work as expected
when connecting to a server that is listening on an encrypted TLS
socket. This test is quite complex as it creates the certificate, sets
up a certificate pool, and explicitly connects via the TLS listen
address with transport credentials.

But ultimately, this complexity is not needed at all as we use
`client.Dial()` to connect to the server, which automatically handles
the setup of the transport credentials when connecting to a TLS-based
socket. And as we set up the `SSL_CERT_FILE` environment variable
already so that it points to the generated certificate, it already gets
picked up by the `client` package via the `x509.SystemCertPool()`.

Simplify the test and remove all the needless setup.