repository: Fix context metadata leaking across testcases
In order to have CreateFork connect to another Gitaly node we need to
write connection information into the gRPC context. In some of our tests
though we leak that information across different testcases. While not
really a problem, it only works correctly by chance.

Fix this by converting the leaking contexts to be testcase-cocal.