catfile: Add encoding information to GitCommit
Currently, Gitaly doesn't return the `encoding` header, which is
populated by `git cat-file` when the commit message was encoded using a
non-UTF-8 encoding and the `i18n.commitEncoding` was set to that
encoding when the commit was made. This can cause clients to show
corrupted commit messages, as seen in [1]. This commit adds the
`encoding` field to the `GitCommit` proto.

1. https://gitlab.com/gitlab-org/gitlab/-/issues/20058

Fixes: https://gitlab.com/gitlab-org/gitaly/-/issues/836
Changelog: added
Signed-off-by: arkn98 <2424696-arkn98@users.noreply.gitlab.com>