proto: Deprecate `root_oid` field in the TreeEntry message
The TreeEntry message we're returning from the `GetTreeEntries()` RPC
contains a `root_oid` field. This field always points to the tree ID of
the commit via which `GetTreeEntries()` arrived at the tree. This design
is quite weird, as the root OID would be the same for every single tree
entry returned by that RPC call. Furthermore, this field requires us to
spawn a separate Git process only to resolve that object ID and thus
slows down overall latency of the RPC.

Deprecate the field so that we can remove it in release 16.0. It is not
known to be used by any clients anyway.

Changelog: deprecated