Makefile: Treat `PROTO_DEST_DIR` as 'new' in diff
For the `no-proto-changes` Makefile target we diff the newly generated
protobufs in `${PROTO_DEST_DIR}` against the in-tree files located in
`proto/go`.

By placing `${PROTO_DEST_DIR}` as the first argument the diff will show
it as the 'old' value, which is unintuitive as it's much more likely to
have been changed. Swap the order so new files are treated as new.