operations: Improve error when formatting tag fails
When writing tags via `localrepo.WriteTag()`, we have a safety check
that checks whether the newly written tag is formatted correctly. The
intent of the check is mostly to verify that it's impossible to inject
tag fields via any of the parameters by checking for additional newlines
that shouldn't actually be there.

In case this check fails we return an error with `Unknown` type and a
Ruby-esque error message. Let's stop doing that though -- the calling
side is never checking for this exact error anyway. And given that this
error condition indicates that the parameters passed by the caller are
invalid, let's also give it a proper error code.

Changelog: changed