repository: Use explicit error code in ApplyGitattributes test
In one of the tests for ApplyGitattributes we're stubbing the
transaction service with a custom voting function. This voting function
is returning a plain error, which ultimately gets converted into an
`Unknown` error. This is something that typically wouldn't happen in
production systems though as all `Unknown` errors returned by gRPC
services would be converted to `Internal`.

Convert the error to instead have an explicit error code. This prepares
for an upcoming change where we modify error code semantics for errors
with an `Unknown` error code.