structerr: Remove `NewUnknown()` constructor
Remove the `NewUnknown()` constructor. There are no users of this
function, and explicitly creating errors with `Unknown` error codes does
not make a lot of sense.

Instead, we're about to change semantics of the `New()` constructor to
return `Unknown` error codes that would always get overridden by more
specific error codes when wrapped.