featureflag: Default `tx_restore_custom_hooks` on
Transactionality is being added to the `SetCustomHooks` RPC via the
`tx_restore_custom_hooks` feature flag. This RPC will be invoked with a
CLI client by GitLab administrators to set custom hooks for a
repository. Since this CLI originates the RPC invocations, feature flag
values are not propagated through the RPC context. This change default
enables the feature flag so tranactions can be enabled for the
`set-hooks` CLI.