gitaly/config: Omit empty fields when serializing as TOML
When serializing fields in the TOML format then we will not omit fields
that are empty. This causes us e.g. convert a `nil` array of values into
an empty array of values, which may or may not have the same semantics.

We will need to really only print what matters in two different
contexts:

    - We will introduce a new command that writes the effective Gitaly
      configuration to its standard output. This can be done to verify
      whether the actual configuration matches your expectations,
      including generated default values. By omitting empty fields here
      we will only print what's relevant.

    - We will introduce tests that need to serialize the data structure
      to a reader in the TOML format and then round-trip it via
      `config.Load()`. This does not currently work as expected due to
      the above conversion.

Add the `omitempty` keyword to the fields' TOML tags. As there is no
code path righgt now that would ever serialize the configuration this
change cannot have any impact on existing code paths. Add a test that
verifies that an empty configuration round-trips as expected now.